Typecasing 0 as int

+++ lib/filecache.php
@@ -497,10 +497,10 @@ class OC_FileCache{
 	public static function triggerUpdate($user=''){
 		if($user) {
 			$query=OC_DB::prepare('UPDATE `*PREFIX*fscache` SET `mtime`=? WHERE `user`=? AND `mimetype`=?');
-			$query->execute(array(0,$user,"httpd/unix-directory"));
+			$query->execute(array((int)0,$user,"httpd/unix-directory"));
 		}else{
 			$query=OC_DB::prepare('UPDATE `*PREFIX*fscache` SET `mtime`=? AND `mimetype`=?');
-			$query->execute(array(0,"httpd/unix-directory"));
+			$query->execute(array((int)0,"httpd/unix-directory"));
 		}
 	}
 }
